@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
 class ConfirmationView extends StatelessWidget {
-  const ConfirmationView({Key key}) : super(key: key);
+  const ConfirmationView({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
